# Add array pooling