--- a/NickBuhro.Translit/CustomStringBuilder.cs
+++ b/NickBuhro.Translit/CustomStringBuilder.cs
@@ -1,13 +1,22 @@
-﻿namespace NickBuhro.Translit
+﻿using System;
+#if NET45 || NETSTANDARD1_3
+using System.Buffers;
+#endif
+
+namespace NickBuhro.Translit
 {
-    internal struct CustomStringBuilder
+    internal struct CustomStringBuilder : IDisposable
     {
         private readonly char[] _array;
         private int _index;
 
         public CustomStringBuilder(int capacity)
         {
+#if NET45 || NETSTANDARD1_3
+            _array = ArrayPool<char>.Shared.Rent(capacity);
+#else
             _array = new char[capacity];
+#endif
             _index = 0;
         }
 
@@ -19,6 +28,13 @@
         public override string ToString()
         {
             return new string(_array, 0, _index);
+        }
+
+        public void Dispose()
+        {
+#if NET45 || NETSTANDARD1_3
+            ArrayPool<char>.Shared.Return(_array);
+#endif
         }
     }
 }