Before:

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.1304 (1803/April2018Update/Redstone4)
Intel Core i5-6200U CPU 2.30GHz (Max: 2.40GHz) (Skylake), 1 CPU, 4 logical and 2 physical cores
Frequency=2343753 Hz, Resolution=426.6661 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4121.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4121.0


```
| Method | Asset |            Mean |            Error |           StdDev |          Median | Scaled | ScaledSD |     Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|------- |------ |----------------:|-----------------:|-----------------:|----------------:|-------:|---------:|----------:|----------:|----------:|-----------:|
| **C2Lv12** |   **Big** |  **8,684,387.2 ns** |   **172,096.759 ns** |   **405,652.309 ns** |  **8,551,224.4 ns** |   **1.00** |     **0.00** |  **203.1250** |  **140.6250** |   **93.7500** |   **798982 B** |
| L2Cv12 |   Big | 64,852,888.6 ns | 1,291,973.854 ns | 1,633,933.253 ns | 64,427,864.2 ns |   7.48 |     0.38 | 5750.0000 | 5750.0000 | 5750.0000 | 22718376 B |
| C2Lv13 |   Big |  8,754,158.1 ns |   144,945.893 ns |   135,582.465 ns |  8,723,302.2 ns |   1.01 |     0.05 |  203.1250 |  187.5000 |  187.5000 |   784880 B |
| L2Cv13 |   Big |  9,587,668.9 ns |   189,538.007 ns |   194,641.560 ns |  9,531,954.5 ns |   1.11 |     0.05 |  187.5000 |  187.5000 |  187.5000 |   751547 B |
| C2Lv14 |   Big |  3,007,367.3 ns |    54,338.051 ns |    50,827.842 ns |  2,996,431.1 ns |   0.35 |     0.02 |  328.1250 |  328.1250 |  328.1250 |  1476064 B |
| L2Cv14 |   Big |  3,258,378.9 ns |    64,932.012 ns |   151,776.484 ns |  3,256,909.2 ns |   0.38 |     0.02 |  195.3125 |  195.3125 |  195.3125 |   750480 B |
|        |       |                 |                  |                  |                 |        |          |           |           |           |            |
| **C2Lv12** | **Small** |      **1,253.0 ns** |        **27.708 ns** |        **69.002 ns** |      **1,223.2 ns** |   **1.00** |     **0.00** |    **0.2842** |         **-** |         **-** |      **449 B** |
| L2Cv12 | Small |      8,694.1 ns |        68.253 ns |        63.844 ns |      8,686.4 ns |   6.96 |     0.35 |    0.9155 |         - |         - |     1444 B |
| C2Lv13 | Small |      1,334.2 ns |        26.644 ns |        29.615 ns |      1,341.1 ns |   1.07 |     0.06 |    0.2499 |         - |         - |      393 B |
| L2Cv13 | Small |      1,316.5 ns |        26.297 ns |        24.599 ns |      1,311.1 ns |   1.05 |     0.06 |    0.1526 |         - |         - |      241 B |
| C2Lv14 | Small |        202.6 ns |         2.477 ns |         2.069 ns |        202.7 ns |   0.16 |     0.01 |    0.2141 |         - |         - |      337 B |
| L2Cv14 | Small |        212.8 ns |         4.165 ns |         5.839 ns |        210.3 ns |   0.17 |     0.01 |    0.1221 |         - |         - |      193 B |

After:

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.1304 (1803/April2018Update/Redstone4)
Intel Core i5-6200U CPU 2.30GHz (Max: 2.40GHz) (Skylake), 1 CPU, 4 logical and 2 physical cores
Frequency=2343753 Hz, Resolution=426.6661 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4121.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4121.0


```
| Method | Asset |            Mean |            Error |           StdDev |          Median | Scaled | ScaledSD |     Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|------- |------ |----------------:|-----------------:|-----------------:|----------------:|-------:|---------:|----------:|----------:|----------:|-----------:|
| **C2Lv12** |   **Big** |  **7,185,680.1 ns** |   **259,381.490 ns** |   **740,029.987 ns** |  **6,941,074.8 ns** |   **1.00** |     **0.00** |  **195.3125** |  **132.8125** |   **78.1250** |   **798992 B** |
| L2Cv12 |   Big | 62,390,638.7 ns | 2,681,055.226 ns | 7,905,151.825 ns | 60,292,616.2 ns |   8.77 |     1.38 | 5222.2222 | 5222.2222 | 5222.2222 | 22718376 B |
| C2Lv13 |   Big |  8,291,180.9 ns |   264,108.197 ns |   778,728.979 ns |  8,047,956.4 ns |   1.16 |     0.15 |  171.8750 |  156.2500 |  156.2500 |   784723 B |
| L2Cv13 |   Big |  8,833,161.3 ns |   366,357.835 ns | 1,074,464.111 ns |  8,448,895.9 ns |   1.24 |     0.19 |  156.2500 |  156.2500 |  156.2500 |   751204 B |
| C2Lv14 |   Big |  2,621,646.8 ns |    51,952.107 ns |   101,328.595 ns |  2,625,750.1 ns |   0.37 |     0.04 |   89.8438 |   89.8438 |   89.8438 |   387664 B |
| L2Cv14 |   Big |  3,296,569.9 ns |    60,513.968 ns |    53,644.051 ns |  3,295,655.8 ns |   0.46 |     0.04 |   85.9375 |   85.9375 |   85.9375 |   362816 B |
|        |       |                 |                  |                  |                 |        |          |           |           |           |            |
| **C2Lv12** | **Small** |      **1,484.5 ns** |        **29.234 ns** |        **28.712 ns** |      **1,483.2 ns** |   **1.00** |     **0.00** |    **0.2842** |         **-** |         **-** |      **449 B** |
| L2Cv12 | Small |     10,995.7 ns |       217.461 ns |       472.742 ns |     10,975.4 ns |   7.41 |     0.34 |    0.9155 |         - |         - |     1444 B |
| C2Lv13 | Small |      1,749.6 ns |        45.280 ns |       132.797 ns |      1,707.2 ns |   1.18 |     0.09 |    0.2499 |         - |         - |      393 B |
| L2Cv13 | Small |      1,413.6 ns |        30.405 ns |        51.629 ns |      1,398.4 ns |   0.95 |     0.04 |    0.1526 |         - |         - |      241 B |
| C2Lv14 | Small |        299.8 ns |         3.571 ns |         3.165 ns |        299.1 ns |   0.20 |     0.00 |    0.0663 |         - |         - |      104 B |
| L2Cv14 | Small |        321.9 ns |         4.444 ns |         3.940 ns |        322.6 ns |   0.22 |     0.00 |    0.0610 |         - |         - |       96 B |