@nick-buhro 2020-06-22
Hi, @sungam3r!

Thanks a lot for your update, it looks interesting. I've never used `ArrayPool` so I considered to research a little and added new benchmark (the code is [here](https://github.com/nick-buhro/Translit/tree/sungam3r-array-pool)). It looks that `ArrayPool` works better on big arrays (1000+ characters), but is worse on small arrays and adds deviation. Will be glad to hear your thoughts:)

P.S.: I'll merge your request now and maybe add option to use GC on small arrays later.

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.18363
Intel Core i5-8250U CPU 1.60GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4180.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4180.0


```
| Method | AssetSize | Direction |            Mean |         Error |        StdDev | Scaled |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------- |---------- |---------- |----------------:|--------------:|--------------:|-------:|---------:|---------:|---------:|----------:|
|     **GC** |        **10** |       **C2L** |        **73.12 ns** |     **0.1155 ns** |     **0.0965 ns** |   **1.00** |   **0.0433** |        **-** |        **-** |     **136 B** |
|     AP |        10 |       C2L |       115.30 ns |     0.1817 ns |     0.1610 ns |   1.58 |   0.0153 |        - |        - |      48 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |        **10** |       **L2C** |        **78.02 ns** |     **0.0729 ns** |     **0.0646 ns** |   **1.00** |   **0.0305** |        **-** |        **-** |      **96 B** |
|     AP |        10 |       L2C |       120.29 ns |     0.0941 ns |     0.0880 ns |   1.54 |   0.0153 |        - |        - |      48 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |       **100** |       **C2L** |       **398.21 ns** |     **0.3251 ns** |     **0.3041 ns** |   **1.00** |   **0.2728** |        **-** |        **-** |     **859 B** |
|     AP |       100 |       C2L |       421.99 ns |     1.0754 ns |     0.8980 ns |   1.06 |   0.0739 |        - |        - |     233 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |       **100** |       **L2C** |       **381.35 ns** |     **0.5601 ns** |     **0.5239 ns** |   **1.00** |   **0.1426** |        **-** |        **-** |     **449 B** |
|     AP |       100 |       L2C |       415.92 ns |     0.3831 ns |     0.3584 ns |   1.09 |   0.0710 |        - |        - |     225 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |      **1000** |       **C2L** |     **7,079.06 ns** |    **43.0747 ns** |    **35.9693 ns** |   **1.00** |   **2.6016** |        **-** |        **-** |    **8213 B** |
|     AP |      1000 |       C2L |     6,868.37 ns |    21.3658 ns |    19.9855 ns |   0.97 |   0.6866 |        - |        - |    2166 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |      **1000** |       **L2C** |     **7,360.41 ns** |    **28.6497 ns** |    **26.7989 ns** |   **1.00** |   **1.2512** |        **-** |        **-** |    **3940 B** |
|     AP |      1000 |       L2C |     7,178.76 ns |    32.7721 ns |    29.0516 ns |   0.98 |   0.6027 |        - |        - |    1910 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |     **10000** |       **C2L** |   **104,137.14 ns** |   **125.2264 ns** |   **111.0100 ns** |   **1.00** |  **25.6348** |        **-** |        **-** |   **81424 B** |
|     AP |     10000 |       C2L |   102,057.14 ns |   152.9323 ns |   135.5705 ns |   0.98 |   6.7139 |        - |        - |   21376 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |     **10000** |       **L2C** |   **104,093.61 ns** |   **243.5574 ns** |   **215.9072 ns** |   **1.00** |  **12.3291** |        **-** |        **-** |   **38872 B** |
|     AP |     10000 |       L2C |   101,764.32 ns |   159.6431 ns |   149.3302 ns |   0.98 |   5.8594 |        - |        - |   18824 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |    **100000** |       **C2L** | **1,143,096.29 ns** | **1,044.1474 ns** |   **925.6094 ns** |   **1.00** | **177.7344** | **177.7344** | **177.7344** |  **813784 B** |
|     AP |    100000 |       C2L | 1,095,623.20 ns | 1,385.6385 ns | 1,228.3323 ns |   0.96 |  48.8281 |  48.8281 |  48.8281 |  213760 B |
|        |           |           |                 |               |               |        |          |          |          |           |
|     **GC** |    **100000** |       **L2C** | **1,127,023.52 ns** |   **426.1020 ns** |   **332.6724 ns** |   **1.00** | **105.4688** | **105.4688** | **105.4688** |  **387320 B** |
|     AP |    100000 |       L2C | 1,084,467.02 ns | 1,959.5126 ns | 1,737.0567 ns |   0.96 |  48.8281 |  48.8281 |  48.8281 |  187296 B |