@sungam3r 2020-07-05
>  It looks that ArrayPool works better on big arrays (1000+ characters), but is worse on small arrays

In terms of performance yes but in terms of memory consumption, absolutely all benchmarks show that it is better.